Update backup.sh
Added checks for (disc)/connected or restricted environments. Install bzip2 package if it is missing in a connected environment. In a disconnected environment the bzip2 package will not be installed and instead the generated backup on the local machine will be in a tar format. If the bzip package is installed on the local machine then the generated tar achive will be further compressed using bzip.